llm-studio/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
pymongo==4.5.0
motor==3.2.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
httpx==0.25.1
python-dotenv==1.0.0
